refactor(idp): split service implementation by responsibility
